refactor directory structure into analytics submodule
